HSEARCH-2481 Make sure SerializationTestHelper test is executed
It used not to be executed in the -engine module, because the test
class didn't end with "Test".
